fix: readonly form versions
