+++ src/app/[locale]/(logged-in)/user/profile/components/ProjectsSafeData/index.ts
@@ -1,0 +1,3 @@
+import ProjectsSafeData from "./ProjectsSafeData";
+
+export default ProjectsSafeData;
