+++ src/app/[locale]/(logged-in)/user/profile/components/ProjectsSafeOutputs/index.ts
@@ -1,0 +1,3 @@
+import ProjectsSafeOutputs from "./ProjectsSafeOutputs";
+
+export default ProjectsSafeOutputs;
