+++ src/app/[locale]/(logged-in)/user/profile/components/ProjectsSafeSettings/index.ts
@@ -1,0 +1,3 @@
+import ProjectsSafeSettings from "./ProjectsSafeSettings";
+
+export default ProjectsSafeSettings;
