+++ src/modules/ProjectsSafeDataForm/index.ts
@@ -1,0 +1,3 @@
+import ProjectsSafeDataForm from "./ProjectsSafeDataForm";
+
+export default ProjectsSafeDataForm;
